docker-compose/etl/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
async-timeout==5.0.1
    # via redis
backoff==2.2.1
    # via -r requirements.in
certifi==2024.8.30
    # via elastic-transport
elastic-transport==8.15.1
    # via elasticsearch
elasticsearch==8.16.0
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
redis==5.2.0
    # via -r requirements.in
urllib3==2.2.3
    # via elastic-transport
pydantic==2.10.2
pydantic-settings==2.6.1